Create Stack and Queue Operations using Push, Pop, Add, Delete in Java
Write a program using JCF (single java file-based project with java file name as linlist) to 
achieve the following: 
Write a program that randomly stores 10 integer numbers (to start with) in a LinkedList 
object. Using this LinkedList object, implement stack and queue operations using GUI as 
follows: Add 2 Radio Buttons for choosing options (a) Stack and (b) Queue, and 4 
JButtons (Push, Pop for Stack, and Add, Delete for Queue).
Once you run the program, it should display a LinkedList of 10 random integers in a text 
field. You should then be able to choose a radio button of either Stack or Queue, 
followed by one of its operations (Push or Pop Buttons for Stack ONLY, and Add or 
Delete Buttons for Queue ONLY (Invalid choices like choosing Stack followed by 
pressing ADD button should not work). The program should then display the modified 
list (as per the operation undertaken) in the text field. 
 
The commands to be entered in showInputDialogue will be as follows: 
Push 
Pop 
Add 
Delete